'Update_Movie'

+++ src/models/admin/list-model/list-movie.service.ts
@@ -1,20 +1,18 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import mongoose, { model } from 'mongoose';
+import mongoose from 'mongoose';
 import { ObjectId } from 'mongodb';
-import { Video, VideoSchema } from 'src/database/schemas/video.schema';
+import { Video } from 'src/database/schemas/video.schema';
 import { Favorites } from 'src/database/schemas/favorite.schema';
 import { Rating } from 'src/database/schemas/rating.schema';
-import { Genre, GenreSchema } from 'src/database/schemas/genre.schema';
+import { Genre } from 'src/database/schemas/genre.schema';
 import MoviesResponseDTO from './dto/MoviesResponse.Dto';
 import MovieDTO from './dto/movie.Dto';
-import { ModelService } from '../../base-repository/Movie/model.service';
 import { MovieRes } from './dto/Movie.model';
-import { Cast, CastSchema } from 'src/database/schemas/cast.schema';
+import { Cast } from 'src/database/schemas/cast.schema';
 import { FireBaseService } from 'src/models/base-repository/firebase/fire-base-service/fire-base-service.service';
-import { Language, LanguageSchema } from 'src/database/schemas/language.schema';
+import { Language } from 'src/database/schemas/language.schema';
 import { JsonResponse } from './json-response.model';
-import { v4 as uuidv4 } from 'uuid';
 import { MESSAGES_CODE } from 'src/Constant/status.constants';
 // import MovieDTO from './Dto/movie.Dto';
 
@@ -34,9 +32,10 @@ export class ListModelService {
     private langugeModel: mongoose.Model<Language>,
     @InjectModel(Cast.name)
     private castModel: mongoose.Model<Cast>,
-    private readonly modelService: ModelService<Video>,
     private readonly firebaseService: FireBaseService,
   ) {}
+
+  // Find All Movie
   async FindAll(
     searchQuery: string | undefined,
     // additionDate: boolean | undefined,
@@ -107,30 +106,154 @@ export class ListModelService {
       };
       listProducts.push(listProduct);
     }
-
     return {
       data: listProducts,
       total: totalPage,
     };
   }
 
-  async addMovie(modelrequest: MovieRes, files: any): Promise<any> {
+  // Update movie
+  async updateAMovie(
+    id: ObjectId,
+    modelRequest: MovieRes,
+    files: any,
+  ): Promise<any> {
+    const jsonResponse = new JsonResponse<Video>(true);
+    if (id === null) {
+      return new NotFoundException(
+        MESSAGES_CODE.UPDATE_FAIL,
+        'id is reqiuired',
+      );
+    }
+    const video = await this.movieModel.findById(id).exec();
+    if (!video) {
+      return new NotFoundException(
+        MESSAGES_CODE.UPDATE_FAIL,
+        'Video not found',
+      );
+    }
+    try {
+      video.title = modelRequest.title;
+      video.duration = modelRequest.duration;
+      video.content = modelRequest.content;
+      video.mpaRatings = modelRequest.mpaRatings;
+      video.country = modelRequest.country;
+      video.releaseYear = modelRequest.releaseYear;
+      video.additionDate = new Date();
+      const casts: Cast[] = [];
+      let count = 0;
+      const existingCast = await this.fetchCast(video.castId);
+      for (const i of modelRequest.cast) {
+        // Check if the cast already exists in the movie
+        const nameCastInFilm = capitalizeFirstLetter(i.nameInFilm);
+        const nameCastOfActor = capitalizeFirstLetter(i.nameOfActor);
+        const existing = existingCast.find(
+          (castId) => castId.nameInFilm === nameCastInFilm,
+        );
+        const avatar = await this.firebaseService.upload(files.avatar[count]);
+        const idAvatar = avatar.result.url;
+
+        if (existing !== undefined) {
+          await this.firebaseService.deleteFile(existing.avatar);
+          // Update existing cast
+          existing.nameOfActor = nameCastOfActor;
+          existing.avatar = idAvatar;
+          casts.push(existing);
+        } else {
+          // Add new cast
+          const castStore = new this.castModel();
+          castStore.nameInFilm = nameCastInFilm;
+          castStore.nameOfActor = nameCastOfActor;
+          castStore.avatar = idAvatar;
+          castStore.videoId = video._id.toString();
+          video.castId.push(castStore._id.toString());
+          casts.push(castStore);
+        }
+        count++;
+      }
+      console.log(video.castId);
+      // Delete casts not present in the update request
+      video.castId.forEach(async (existingCast) => {
+        const castExistsInUpdate = casts.some(
+          (updatedCast) =>
+            updatedCast._id.toString() === existingCast.toString(),
+        );
+        if (!castExistsInUpdate) {
+          console.log('existingCast', existingCast);
+          const objectId = new (ObjectId as any)(existingCast);
+          console.log('castExistsInUpdate', castExistsInUpdate);
+          // Delete the cast that is not present in the update request
+          const valueDelete = await this.castModel
+            .findByIdAndDelete(objectId)
+            .exec();
+          this.firebaseService.deleteFile(valueDelete.avatar);
+          video.castId.splice(video.castId.indexOf(existingCast), 1);
+        }
+      });
+      console.log(video.castId);
+      const genres: Genre[] = [];
+      for (const i of modelRequest.genre) {
+        const nameGenre = capitalizeFirstLetter(i.name);
+        const existGenre = await this.genreModel
+          .findOne({ name: nameGenre })
+          .exec();
+        if (!existGenre) {
+          const genreStore = new this.genreModel();
+          genreStore.name = nameGenre;
+          genres.push(genreStore);
+          video.genreId.push(genreStore._id.toString());
+        }
+      }
+      const languages: Language[] = [];
+      for (const i of modelRequest.language) {
+        const nameLanguage = capitalizeFirstLetter(i.title);
+        const existLanguage = await this.langugeModel
+          .findOne({ title: nameLanguage })
+          .exec();
+        if (!existLanguage) {
+          const languageStore = new this.langugeModel();
+          languageStore.title = nameLanguage;
+          languages.push(languageStore);
+          video.languageId.push(languageStore._id.toString());
+        }
+      }
+      await this.firebaseService.updateFile(
+        video.posterImage,
+        files.posterImage[0],
+      );
+      await this.firebaseService.updateFile(video.movieLink, files.movieUrl[0]);
+      // Insert data
+      console.log(video.castId);
+      await this.movieModel.updateMany({ _id: id }, video);
+      await this.castModel.create(casts);
+      await this.genreModel.create(genres);
+      await this.langugeModel.create(languages);
+      jsonResponse.result = video;
+      jsonResponse.message = MESSAGES_CODE.UPDATED_SUCCESS;
+      return { jsonResponse };
+    } catch (error) {
+      throw new Error(error);
+    }
+  }
+
+  // Add new movie
+  async addMovie(modelRequest: MovieRes, files: any): Promise<any> {
     const jsonResponse = new JsonResponse<Video>(true);
     try {
       const video = new this.movieModel({
-        title: modelrequest.title,
-        duration: modelrequest.duration,
-        content: modelrequest.content,
-        mpaRatings: modelrequest.mpaRatings,
-        country: modelrequest.country,
-        releaseYear: modelrequest.releaseYear,
+        title: modelRequest.title,
+        duration: modelRequest.duration,
+        content: modelRequest.content,
+        mpaRatings: modelRequest.mpaRatings,
+        country: modelRequest.country,
+        releaseYear: modelRequest.releaseYear,
         additionDate: new Date(),
         view: 0,
       });
 
       const casts: Cast[] = [];
       let count = 0;
-      for (const i of modelrequest.cast) {
+      for (const i of modelRequest.cast) {
         const nameCastInFilm = capitalizeFirstLetter(i.nameInFilm);
         const nameCastOfActor = capitalizeFirstLetter(i.nameOfActor);
         const avatar = await this.firebaseService.upload(files.avatar[count]);
@@ -145,13 +268,12 @@ export class ListModelService {
         count++;
       }
       const genres: Genre[] = [];
-      for (const i of modelrequest.genre) {
+      for (const i of modelRequest.genre) {
         const nameGenre = capitalizeFirstLetter(i.name);
         const existGenre = await this.genreModel
           .findOne({ name: nameGenre })
           .exec();
         if (existGenre) {
-          console.log(1);
           video.genreId.push((await existGenre)._id.toString());
         } else {
           const genreStore = new this.genreModel();
@@ -161,13 +283,12 @@ export class ListModelService {
         }
       }
       const languages: Language[] = [];
-      for (const i of modelrequest.language) {
+      for (const i of modelRequest.language) {
         const nameLanguage = capitalizeFirstLetter(i.title);
         const existLanguage = await this.langugeModel
           .findOne({ title: nameLanguage })
           .exec();
         if (existLanguage) {
-          console.log(2);
           video.languageId.push((await existLanguage)._id.toString());
         } else {
           const languageStore = new this.langugeModel();
@@ -176,13 +297,11 @@ export class ListModelService {
           video.languageId.push(languageStore._id.toString());
         }
       }
-
       const idImage = await this.firebaseService.upload(files.posterImage[0]);
       const idMovie = await this.firebaseService.upload(files.movieUrl[0]);
       video.posterImage = await idImage.result.url;
       video.movieLink = await idMovie.result.url;
       // Insert data
-      await this.movieModel.create(video);
       await this.movieModel.create(video);
       await this.castModel.create(casts);
       await this.genreModel.create(genres);
@@ -198,32 +317,50 @@ export class ListModelService {
     }
   }
 
+  // Delete a movie
   async deleteAMovie(id: ObjectId): Promise<any> {
     if (id === null) {
-      throw new Error(MESSAGES_CODE.CREATE_FAIL);
+      return new NotFoundException(
+        MESSAGES_CODE.DELETE_FAIL,
+        'Video not found',
+      );
     }
-    const video = await this.modelService.deleteMovie(id);
-    await this.firebaseService.deleteFile(video.posterImage);
-    await this.firebaseService.deleteFile(video.movieLink);
-    const videoId = video._id.toString();
-    const cast = await this.castModel.find({ videoId: videoId }).exec();
-    if (cast.length > 0) {
-      for (const i of cast) {
-        await this.castModel.findByIdAndDelete(i._id).exec();
-        await this.firebaseService.deleteFile(i.avatar);
+    const value = await this.movieModel.findById(id).exec();
+    if (!value) {
+      return new NotFoundException(
+        MESSAGES_CODE.DELETE_FAIL,
+        'Video not found',
+      );
+    }
+    const deleteMovie = await this.movieModel.findByIdAndDelete(id).exec();
+    if (deleteMovie !== null) {
+      await this.firebaseService.deleteFile(deleteMovie.posterImage);
+      await this.firebaseService.deleteFile(deleteMovie.movieLink);
+      const videoId = deleteMovie._id.toString();
+      const cast = await this.castModel.find({ videoId: videoId }).exec();
+      if (cast.length > 0) {
+        for (const i of cast) {
+          await this.castModel.findByIdAndDelete(i._id).exec();
+          await this.firebaseService.deleteFile(i.avatar);
+        }
       }
     }
-    return video;
+    return MESSAGES_CODE.DELETED_SUCCESS;
   }
 
-  async deleteListMovie(idList: ObjectId[]): Promise<void> {
+  // Delete list movie
+  async deleteListMovie(idList: ObjectId[]): Promise<any> {
     if (idList.length === 0) {
-      throw new Error('Video not found');
+      return new NotFoundException(
+        MESSAGES_CODE.DELETE_FAIL,
+        'Video not found',
+      );
     }
     let video;
     for (const id of idList) {
       video = this.deleteAMovie(id);
     }
+
     return video;
   }
 
@@ -250,6 +387,24 @@ export class ListModelService {
     }
   }
 
+  private async fetchCast(cast: string[]): Promise<Cast[]> {
+    // Assume that you have a "Favorites" model/schema defined in your application
+    // and a corresponding collection in MongoDB
+    try {
+      const listcast: Cast[] = [];
+      // Assuming you have a GenreModel or a similar model
+      for (const models of cast) {
+        const casts = await this.castModel.findById({ _id: models }).exec();
+        listcast.push(casts);
+      }
+      return listcast;
+
+      // Map the genre data to the desired format
+    } catch (error) {
+      // Handle any errors that might occur during the database query
+      throw new Error('Error fetching favorites');
+    }
+  }
   private async fetchGenres(genre: string[]): Promise<Genre[]> {
     // Assume that you have a "Favorites" model/schema defined in your application
     // and a corresponding collection in MongoDB
